requirements.txt
gitdb==4.0.11
GitPython==3.1.43
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
rich==13.7.1
smmap==5.0.1
